+++ apps/payload/src/payload-types.ts
@@ -666,12 +666,20 @@ export interface HeroBlock {
   opacity?: number | null;
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -699,12 +707,20 @@ export interface TextSectionBlock {
   };
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -735,12 +751,20 @@ export interface ContentBlock {
   };
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -773,12 +797,20 @@ export interface FaqBlock {
   }[];
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -805,12 +837,20 @@ export interface TestimonialsListBlock {
   showAvatar?: boolean | null;
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -871,12 +911,20 @@ export interface CardsGridBlock {
   }[];
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -927,12 +975,20 @@ export interface CarouselBlock {
   }[];
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -969,12 +1025,20 @@ export interface LogosBlock {
   }[];
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -1011,12 +1075,20 @@ export interface LinksListBlock {
   }[];
   section?: {
     theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-    marginTop?: ('none' | 'base' | 'large') | null;
-    marginBottom?: ('none' | 'base' | 'large') | null;
-    paddingX?: ('none' | 'base' | 'large') | null;
     paddingY?: ('none' | 'base' | 'large') | null;
-    maxWidth?: ('none' | 'base' | 'small') | null;
-    backgroundImage?: (number | null) | Media;
+    paddingX?: ('none' | 'base') | null;
+    maxWidth?: ('none' | 'base') | null;
+    background?: {
+      /**
+       * Upload an image or video. Use the "Background" folder.
+       */
+      media?: (number | null) | Media;
+      overlay?: ('black' | 'white') | null;
+      /**
+       * 0 = transparent, 100 = fully opaque
+       */
+      opacity?: number | null;
+    };
   };
   id?: string | null;
   blockName?: string | null;
@@ -1134,12 +1206,20 @@ export interface Preset {
         opacity?: number | null;
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1163,12 +1243,20 @@ export interface Preset {
         };
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1195,12 +1283,20 @@ export interface Preset {
         };
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1229,12 +1325,20 @@ export interface Preset {
         }[];
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1257,12 +1361,20 @@ export interface Preset {
         showAvatar?: boolean | null;
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1304,12 +1416,20 @@ export interface Preset {
         }[];
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1356,12 +1476,20 @@ export interface Preset {
         }[];
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1394,12 +1522,20 @@ export interface Preset {
         }[];
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -1432,12 +1568,20 @@ export interface Preset {
         }[];
         section?: {
           theme?: ('light' | 'dark' | 'light-gray' | 'dark-gray') | null;
-          marginTop?: ('none' | 'base' | 'large') | null;
-          marginBottom?: ('none' | 'base' | 'large') | null;
-          paddingX?: ('none' | 'base' | 'large') | null;
           paddingY?: ('none' | 'base' | 'large') | null;
-          maxWidth?: ('none' | 'base' | 'small') | null;
-          backgroundImage?: (number | null) | Media;
+          paddingX?: ('none' | 'base') | null;
+          maxWidth?: ('none' | 'base') | null;
+          background?: {
+            /**
+             * Upload an image or video. Use the "Background" folder.
+             */
+            media?: (number | null) | Media;
+            overlay?: ('black' | 'white') | null;
+            /**
+             * 0 = transparent, 100 = fully opaque
+             */
+            opacity?: number | null;
+          };
         };
         id?: string | null;
         blockName?: string | null;
@@ -2036,12 +2180,16 @@ export interface HeroBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2056,12 +2204,16 @@ export interface TextSectionBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2079,12 +2231,16 @@ export interface ContentBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2106,12 +2262,16 @@ export interface FaqBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2136,12 +2296,16 @@ export interface TestimonialsListBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2183,12 +2347,16 @@ export interface CardsGridBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2216,12 +2384,16 @@ export interface CarouselBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2257,12 +2429,16 @@ export interface LogosBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2293,12 +2469,16 @@ export interface LinksListBlockSelect<T extends boolean = true> {
     | T
     | {
         theme?: T;
-        marginTop?: T;
-        marginBottom?: T;
-        paddingX?: T;
         paddingY?: T;
+        paddingX?: T;
         maxWidth?: T;
-        backgroundImage?: T;
+        background?:
+          | T
+          | {
+              media?: T;
+              overlay?: T;
+              opacity?: T;
+            };
       };
   id?: T;
   blockName?: T;
@@ -2487,12 +2667,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2505,12 +2689,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2526,12 +2714,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2551,12 +2743,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2579,12 +2775,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2624,12 +2824,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2655,12 +2859,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2694,12 +2902,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;
@@ -2728,12 +2940,16 @@ export interface PresetsSelect<T extends boolean = true> {
                 | T
                 | {
                     theme?: T;
-                    marginTop?: T;
-                    marginBottom?: T;
-                    paddingX?: T;
                     paddingY?: T;
+                    paddingX?: T;
                     maxWidth?: T;
-                    backgroundImage?: T;
+                    background?:
+                      | T
+                      | {
+                          media?: T;
+                          overlay?: T;
+                          opacity?: T;
+                        };
                   };
               id?: T;
               blockName?: T;