+++ apps/storyblok/src/components/Header/index.tsx
@@ -7,7 +7,6 @@ import { Header as HeaderUI } from "@shared/ui";
 
 import { prepareImageProps } from "@/lib/adapters/prepareImageProps";
 import { prepareLinkProps } from "@/lib/adapters/prepareLinkProps";
-import SectionContainer from "@/components/SectionContainer";
 
 import type { IHeaderProps } from "./types";
 
@@ -18,22 +17,10 @@ export default function Header({ blok }: IHeaderProps) {
     return <EmptyBlock name={blok.component as string} />;
 
   return (
-    <SectionContainer
-      blok={{
-        ...blok,
-        paddingY: "none",
-        paddingX: "none",
-        marginTop: "none",
-        marginBottom: "none",
-        theme: undefined,
-      }}
-      className="sticky left-0 top-0 z-50 rounded-t-none"
-    >
-      <HeaderUI
-        links={links.map(prepareLinkProps)}
-        image={prepareImageProps(image[0])}
-        alignVariant={alignVariant as AlignVariant}
-      />
-    </SectionContainer>
+    <HeaderUI
+      links={links.map(prepareLinkProps)}
+      image={prepareImageProps(image[0])}
+      alignVariant={alignVariant as AlignVariant}
+    />
   );
 }