+++ apps/storyblok/src/contentSections/CardsGrid/index.tsx
@@ -1,6 +1,9 @@
-import EmptyBlock from "@shared/ui/components/EmptyBlock";
+import { storyblokEditable } from "@storyblok/react/rsc";
 
+import EmptyBlock from "@shared/ui/components/EmptyBlock";
 import { CardsGrid as CardsGridUI } from "@shared/ui";
+
+import type { IDefaultCardProps } from "@shared/ui/components/sections/cardsGrid/types";
 
 import { prepareImageProps } from "@/lib/adapters/prepareImageProps";
 import { prepareLinkProps } from "@/lib/adapters/prepareLinkProps";
@@ -9,20 +12,27 @@ import SectionContainer from "@/components/SectionContainer";
 import type { ICardsGridProps } from "./types";
 
 export default function CardsGrid({ blok }: ICardsGridProps) {
-  const { items, columns } = blok;
+  const { items, columns, section, _uid } = blok;
 
-  if (items.length === 0) return <EmptyBlock name={blok.component as string} />;
+  if (!items?.length) return <EmptyBlock name={blok.component as string} />;
 
-  const formattedItems = items.map((item) => ({
-    ...item,
-    type: item._type,
+  const formattedItems: IDefaultCardProps[] = items.map((item) => ({
+    title: item.title,
+    description: item.description,
     image: prepareImageProps(item?.image?.[0]),
     link: prepareLinkProps(item?.link?.[0]),
+    alignVariant: (item.alignVariant || "left") as IDefaultCardProps["alignVariant"],
+    rounded: (item.rounded || "none") as IDefaultCardProps["rounded"],
+    backgroundColor: (item.backgroundColor || "none") as IDefaultCardProps["backgroundColor"],
   }));
 
   return (
-    <SectionContainer blok={blok}>
-      <CardsGridUI items={formattedItems} columns={parseInt(columns)} />
+    <SectionContainer
+      sectionData={section?.[0]}
+      id={_uid}
+      editableAttrs={storyblokEditable(blok)}
+    >
+      <CardsGridUI items={formattedItems} columns={parseInt(columns || "3")} />
     </SectionContainer>
   );
 }